POSIX C99 uint32_t for images
It appears that this type is not guaranteed to be available when using
C99. POSIX appears to demand it. If you're trying to compile on a
platform without this type being available, please let me know.

Either way, this commit improves the situation. We need exactly 32 bits.
I'd rather have the build fail than silently ignore an error.