TODO: Move EWMH TODOs to backlog
These two parts of the protocol are not mission critical, I think. I
have not yet come across a client that relies on them. Of course,
there's tools like wmctrl, but I don't use these, so this would result
in a lot of unmaintained and untested code. I'd rather avoid that,
unless a really good reason to implement those pops up.