Focus windows before moving them
It appears to be important in which order EnterNotify events and FocusIn
events are delivered to applications. In EnterNotify events, there's a
boolean field "focus", which is set depending on whether the window
already has focus.

It also appears that some applications need an EnterNotify event with
"focus YES" *after* they get a FocusIn event. Other applications don't
like "focus YES" immediately followed by a FocusOut event.

None of this is well documented in X11 (or I was too stupid to find it).
This commit is experimental and needs testing.