Focus our dummy window in setup()
Scenario: The user issues "startx", katriawm launches from .xinitrc, no
windows are mapped before katriawm has started. Without this commit, the
root window now has focus and, thus, we're in a sloppy-ish mode.

If the user now starts an application which spawns right under the mouse
cursor, it'll get an EnterNotify event with "focus YES". This might
sound reasonable, because that window will indeed get focused. But as
experiments have shown (also see previous commits on this topic), we
should try hard to not generate these kinds of events.

With this commit, we will focus our dummy window as soon as possible.
This will greatly reduce the risk of accidentally creating such an
event.

Again, this is almost exclusively relevant for VTE applications.