Rename variable in manage_focus_remove()
"new_focus" refers to the variable in manage_focus_set(). It's a client
that might be in the list, will be removed from the list, and then be
put at the head of the list. Okay. But manage_focus_remove() is a
generic function that can remove any client, so the whole "new_focus"
business is just not apropriate.