Handle XRandR reporting no monitors
This bug has been present for a long time. It has now shown up due to
the assertion of "nmon > 0". We previously checked for "sr->ncrtc > 0"
which would only happen if there were no *physical* outputs. Now,
"nmon > 0" has checked for *active* outputs.

Depending on the tool you use to configure your screen, it *can* happen
that XRandR reports no active outputs. It's valid and we must handle it.

In theory, it was possible to trigger a crash here if you sent some
appropriate event to the WM while "monitors" was NULL. (It never crashed
on my in real life, though, which is one of the reasons why it took so
long to find this bug.)